moved \Drupal\Core\Response\ResourceResponseTrait
