+++ core/modules/rest/src/ResourceResponse.php
@@ -5,6 +5,7 @@ namespace Drupal\rest;
 use Drupal\Core\Cache\CacheableResponseInterface;
 use Drupal\Core\Cache\CacheableResponseTrait;
 use Drupal\Core\Response\ResourceResponseInterface;
+use Drupal\Core\Response\ResourceResponseTrait;
 use Symfony\Component\HttpFoundation\Response;
 
 /**
